Add php-dom extension for CentOs platform.
The build fails on my CentoOs6.3 base box because drush install throws
an Exception when it does not find the php-dom extension.

The minimum PHP package on CentOs does not include the php-dom extension by default.

Note there is no php:module_dom recipe to add via the requirement stament.  Also, the
maintainer of the php cookbook has depricated all the module_<extension>
recipes, and is asking developers that require these extensions to include them via a call to package as I do here.