Check for a default site before disabling it.
The build fails on our centos63 platform because the default site is not
enabled and the `a2dissite` command throws an error when it does not find it.
This commit corrects this issue by making sure that command is only envoked
when a default site is present.

Note that in the current apache2 cookbook `default['apache']['default_site_enabled']`
is set to `false` for all platforms.