Add iptables configuration for CentOs.
The CentOS VM firewall is very restrictive out of the box. This commit
uses iptables to reconfigure the firewall so that ports 80, 443, and 22
are open for business.  This is the last step needed to make CentOs
boxes work with the drupal-cookbook.

With regard to the firewall recipe that comes with this cookbook:
Unfortunately it does not work for CentOS because CentOS useses iptables
for the firewall.  The README for the iptables cookbook mentions that
there was an effort to merge iptables into the firewall cookbook as an LWRP.
However, as of the time of this commit nothing has come of that effort.